Refactor cli tree
